fix: 생성자 순환참조 이슈로 인한 수정
