+++ src/main/java/com/ll/medium_mission/token/service/TokenService.java
@@ -1,11 +1,16 @@
 package com.ll.medium_mission.token.service;
 
+import com.ll.medium_mission.global.provider.JwtTokenProvider;
+import com.ll.medium_mission.member.service.MemberService;
 import com.ll.medium_mission.token.entity.Token;
 import com.ll.medium_mission.token.repository.TokenRepository;
 import jakarta.transaction.Transactional;
+import java.util.List;
 import java.util.Optional;
+import javax.security.sasl.AuthenticationException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 @Service
 @RequiredArgsConstructor
